Add : Set favorite code

+++ src/services/favorite.service.ts
@@ -2,8 +2,6 @@ import { Injectable } from '@nestjs/common';
 import Service from './service';
 import FavoriteRepository from '../repositories/favorite.repository';
 import { FavoriteInputType, FavoriteType } from '../types/favorite';
-import Favorite from '../models/favorite.entity';
-//import Favorite from '../models/favorite.entity';
 
 @Injectable()
 export default class FavoriteService extends Service {
@@ -13,7 +11,6 @@ export default class FavoriteService extends Service {
 
   async setFavoriteStore(options?: FavoriteInputType): Promise<number> {
     const countQueryBuilder = this.favoriteRepository.createQueryBuilder('favorite');
-    const setQueryBuilder = this.favoriteRepository.createQueryBuilder();
 
     countQueryBuilder.andWhere('userId = :userId', { userId: options.userId });
     countQueryBuilder.andWhere('storeId = :storeId', { storeId: options.storeId });
@@ -23,25 +20,14 @@ export default class FavoriteService extends Service {
     const result = await countQueryBuilder.getCount();
 
     if (result > 0) {
-      setQueryBuilder
-        .delete()
-        .from(Favorite)
-        .where('userId = :userId', { userId: options.userId })
-        .andWhere('storeId = :storeId', { storeId: options.storeId })
-        .execute();
+      this.favoriteRepository.delete({ userId: options.userId, storeId: options.storeId });
     } else {
       const userFavorite = new FavoriteType();
       userFavorite.userId = options.userId;
       userFavorite.storeId = options.storeId;
 
-      setQueryBuilder
-        .insert()
-        .values(userFavorite)
-        .execute();
+      this.favoriteRepository.insert(userFavorite);
     }
-
-    console.log(countQueryBuilder.getQueryAndParameters());
-    console.log(setQueryBuilder.getQueryAndParameters());
 
     return result;
   }
